+++ src/python/pants/base/BUILD
@@ -3,4 +3,42 @@
 
 python_library()
 
-python_tests(name='tests')
+python_tests(
+  name="tests",
+  sources=['*_test.py', '!exception_sink_test.py', '!*_integration_test.py'],
+  dependencies = [
+    '//:build_root',
+  ],
+)
+
+python_tests(
+  name='exception_sink',
+  sources=['exception_sink_test.py'],
+  tags=['platform_specific_behavior'],
+  timeout=90,
+)
+
+python_integration_tests(
+  name='exception_sink_integration',
+  sources=['exception_sink_integration_test.py'],
+  dependencies=[
+    'examples/src/python/example:hello_directory',
+    'testprojects/src/python:coordinated_runs_directory',
+    'testprojects:pants_plugins_directory',
+  ],
+  uses_pants_run=True,
+  tags=['platform_specific_behavior'],
+  # NB: This frequently times out, but due to hanging. So, we want to fail eagerly. See
+  # https://github.com/pantsbuild/pants/issues/8127.
+  timeout=200,
+)
+
+python_integration_tests(
+  name='exiter_integration',
+  sources=['exiter_integration_test.py'],
+  dependencies=[
+    'testprojects/src/python:unicode_directory',
+  ],
+  uses_pants_run=True,
+  timeout=120,
+)
