+++ src/python/pants/base/deprecated_test.py
