+++ src/python/pants/base/exception_sink_integration_test.py
