+++ src/python/pants/base/exception_sink_test.py
