+++ src/python/pants/base/exiter_integration_test.py
