+++ src/python/pants/base/hash_utils_test.py
